Fix ntpfrob/bumpclock to work with large values.

+++ ntpfrob/ntpfrob.h
@@ -2,11 +2,12 @@
  * Our methods, one per linked module
  */
 #include <stdbool.h>      /* for bool */
+#include <stdint.h>       /* for int64_t */
 #include "ntp_fp.h"       /* for l_fp */
 
 typedef enum {plain_text, raw, json} iomode;
 
-extern void bumpclock(int);
+extern void bumpclock(int64_t);
 extern void get_clocktime(l_fp *now);
 extern void jitter(const iomode mode) __attribute__((noreturn));
 extern void ppscheck(const char *device) __attribute__((noreturn));
